Split converter into convert sub module
